added drinks section

+++ app/src/main/java/com/example/neworderingsystem/FoodData.kt
@@ -2,4 +2,4 @@ package com.example.neworderingsystem
 
 import java.math.BigDecimal
 
-data class FoodData(val image: Int, val name: String, val price: Double, var check: Boolean)
+data class FoodData(val image: Int, val name: String, val price: Int, var check: Boolean)
